feat!: remove `dic` from options of setup
