+++ after/plugin/cmp_dictionary.vim
@@ -3,7 +3,7 @@ if exists("g:loaded_cmp_dictionary")
 endif
 let g:loaded_cmp_dictionary = 1
 
-lua require("cmp").register_source("dictionary", require("cmp_dictionary").new())
+lua require("cmp").register_source("dictionary", require("cmp_dictionary.source").new())
 
 command CmpDictionaryUpdate :lua require("cmp_dictionary").update()
